+++ tests/lib/js/static_interop_test/js_types_test.dart
@@ -1007,6 +1007,13 @@ Future<void> asyncTests() async {
     Expect.equals('resolved', (await f).toDart);
   }
 
+  {
+    final f = JSPromise.resolve('resolved'.toJS).toDart;
+    Expect.equals('resolved', (await f).toDart);
+  }
+
+  Expect.throws(() => JSPromise.resolve(JSPromise.resolve('resolved'.toJS)));
+
   // Test resolution with incorrect type.
   // TODO(54214): This type error is not caught in the JS compilers correctly.
   // {
@@ -1038,6 +1045,16 @@ Future<void> asyncTests() async {
   {
     try {
       await getRejectedPromise<JSString>().toDart;
+      Expect.fail('Expected rejected promise to throw.');
+    } catch (e) {
+      final jsError = e as JSObject;
+      Expect.equals('Error: rejected', jsError.toString());
+    }
+  }
+
+  {
+    try {
+      await JSPromise.reject(JSError('rejected')).toDart;
       Expect.fail('Expected rejected promise to throw.');
     } catch (e) {
       final jsError = e as JSObject;