+++ CHANGELOG.md
@@ -55,6 +55,8 @@
 - Added the `JSArray.fromAsync` static method to create an array from an
   asynchronous generator or a synchronous list of asynchronous values.
 
+- Added `JSPromise.resolve` and `JSPromise.reject` static methods.
+
 [#61353]: https://github.com/dart-lang/sdk/issues/61353
 [#62699]: https://github.com/dart-lang/sdk/issues/62699